LayoutTests/editing/pasteboard/paste-blockquote-into-blockquote-3.html

<!DOCTYPE html>
<html>
<head>
<script src="../../resources/dump-as-markup.js"></script>
<style>
    blockquote {
        color: blue;
        border-left: 2px solid blue;
        margin: 0px;
        padding: 0 0 0 20px;
    }
</style>
</head>
<body contenteditable>
<div>
<p>This tests pasting a blockquote into a blockquote. Nothing below should be double-blockquoted.</p>
    <div>
        <blockquote type="cite">one<br>two<br>three<br id="endContent"><br><br></blockquote>
    </div>
    <br>
</div>
<script>

var sel = window.getSelection();
var endContent = document.getElementById("endContent");
var range = new Range;
range.selectNode(endContent);
sel.setPosition(range.endContainer, range.endOffset);
document.execCommand("InsertHTML", false, "<blockquote type='cite'><div>two</div></blockquote>");

Markup.description(document.getElementsByTagName('p')[0].innerText);
Markup.dump(document.getElementsByTagName('div')[1]);

</script>
</body>
</html>
