LayoutTests/editing/pasteboard/paste-blockquote-into-blockquote-3-expected.txt

This tests pasting a blockquote into a blockquote. Nothing below should be double-blockquoted.
| "\n        "
| <blockquote>
|   type="cite"
|   "one"
|   <br>
|   "two"
|   <br>
|   "three"
|   <br>
|     id="endContent"
|   "two<#selection-caret>"
|   <br>
| <blockquote>
|   type="cite"
|   <br>
| "\n    "
